+++ crates/cli/utils/src/load_secret_key.rs
